+++ CHANGES.rst
@@ -9,6 +9,17 @@ CHANGES
 
 in-progress
 ===========
+
+
+2017-03-17 0.13.0
+=================
+- Add x-www-form-urlencoded serialization to :ref:`lib-terkin-data` library
+- Fix JSON serialization: Don’t convert values to string
+- Improve CSV serialization demo: Set float precision to three digits
+
+
+2017-02-01 unreleased
+=====================
 - Add JSON serialization to :ref:`lib-terkin-data` library